rag_conversation/packages/rag-conversation/pyproject.toml
[tool.poetry]
name = "rag-conversation"
version = "0.0.1"
description = "RAG conversation"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
openai = "<2"
langchain-community = ">=0.0.7,<0.2"
langchain-core = ">=0.1.4,<0.2"

[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.21"
fastapi = ">=0.104.0,<1"
sse-starlette = "^1.6.5"

[tool.langserve]
export_module = "rag_conversation.chain"
export_attr = "chain"

[tool.templates-hub]
use-case = "chatbot"
author = "LangChain"
integrations = ["OpenAI"]
tags = ["getting-started"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
